Adding register configuration to MCAL driver

+++ SAMV71x/app/08_UartBase/src/Bsw/Mcal/Com/Uart/include/Uart_Driver.h
@@ -14,7 +14,7 @@
 /* ************************************************************************** */
 /* Public function prototypes  
 /* ************************************************************************** */
-void UART_Driver_Configure( Uart *uart, uint8_t mode, uint8_t clkSrc, uint8_t parity, uint32_t baudrate, uint32_t clkSrcFreq );
+void UART_Driver_Configure( Uart *uart, uint16_t mode, uint16_t clkSrc, uint16_t parity, uint32_t baudrate, uint32_t clkSrcFreq );
 void UART_Driver_SetBaudrate( Uart *uart, uint32_t baudrate, uint32_t clkSrcFreq );
 void UART_Driver_SetIt( Uart *uart, uint8_t enable, uint32_t mode );
 void UART_Driver_SetTx( Uart *uart, uint8_t enable );
